re-introduce ActionView::Component::TestHelpers
This module should not have been removed in #239, as doing so
was a breaking change.

This change re-introduces the module and moves render_component,
which will be removed in v2.0.0, to the module.

Thanks to @jcoyne for the report!

[fixes: #254]
