+++ lib/view_component/test_helpers.rb
@@ -24,14 +24,6 @@ module ViewComponent
       @request ||= ActionDispatch::TestRequest.create
     end
 
-    def render_component(component, **args, &block)
-      ActiveSupport::Deprecation.warn(
-        "`render_component` has been deprecated in favor of `render_inline`, and will be removed in v2.0.0."
-      )
-
-      render_inline(component, args, &block)
-    end
-
     def with_variant(variant)
       old_variants = controller.view_context.lookup_context.variants
